+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## 0.5.3-dev4
+## 0.5.3-dev5
 
 ### Enhancements
 
+* `auto.partition()` can now load Unstructured ISD json documents.
 * Simplify partitioning functions.
 * Improve logging for ingest CLI.
 
@@ -11,6 +12,7 @@
   to pages with similar names.
 * Add optional `encoding` argument to the `partition_(text/email/html)` functions.
 * Added Google Drive connector for ingest cli.
+* Added Gitlab connector for ingest cli.
 
 ### Fixes
 
@@ -280,4 +282,3 @@ of an email.
 ## 0.2.0
 
 * Initial release of unstructured
-